style: reorganized SecurityConfig test and HomeController test
