+++ src/test/java/com/poolc/springproject/poolcreborn/controller/HomeControllerTest.java
@@ -1,10 +1,12 @@
 package com.poolc.springproject.poolcreborn.controller;
 
+import org.junit.jupiter.api.DisplayName;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.test.context.support.WithAnonymousUser;
+import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.junit.Test;
@@ -22,11 +24,20 @@ public class HomeControllerTest {
     private MockMvc mockMvc;
 
     @Test
+    @DisplayName("익명으로 홈페이지 접근")
     @WithAnonymousUser
-    public void 메인화면() throws Exception{
+    public void index_anonymous() throws Exception {
         mockMvc.perform(get("/"))
                 .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(view().name("index"));
+                .andExpect(status().isOk());
+    }
+
+    @Test
+    @DisplayName("유저로 홈페이지 접근")
+    @WithMockUser
+    public void index_user() throws Exception {
+        mockMvc.perform(get("/"))
+                .andDo(print())
+                .andExpect(status().isOk());
     }
 }
