--- src/test/java/com/poolc/springproject/poolcreborn/security/SecurityConfigTest.java
@@ -1,43 +1,0 @@
-package com.poolc.springproject.poolcreborn.security;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.security.test.context.support.WithAnonymousUser;
-import org.springframework.security.test.context.support.WithMockUser;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.junit.Test;
-import org.springframework.test.web.servlet.MockMvc;
-
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@RunWith(SpringRunner.class)
-@AutoConfigureMockMvc
-public class SecurityConfigTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Test
-    @DisplayName("익명으로 홈페이지 접근")
-    @WithAnonymousUser
-    public void index_anonymous() throws Exception {
-        mockMvc.perform(get("/"))
-                .andDo(print())
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    @DisplayName("유저로 홈페이지 접근")
-    @WithMockUser
-    public void index_user() throws Exception {
-        mockMvc.perform(get("/"))
-                .andDo(print())
-                .andExpect(status().isOk());
-    }
-
-
-}
